# docs/swagger.yaml
basePath: /
definitions:
  database.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Простая документация для сервиса подписок
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Subscription'
            type: array
      summary: Get all subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription
      parameters:
      - description: Subscription
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/database.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Subscription'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscription
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      description: Calculate total cost of subscriptions for a period
      parameters:
      - description: Start period (MM-YYYY)
        in: query
        name: period_start
        required: true
        type: string
      - description: End period (MM-YYYY)
        in: query
        name: period_end
        required: true
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      - description: User name
        in: query
        name: user_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get total subscription cost
      tags:
      - subscriptions
swagger: "2.0"